+++ checkout-ui-custom/src/controller/ViewController.js
@@ -29,37 +29,31 @@ const ViewController = (() => {
     showMixedProductsMsg: false
   };
 
-  const config = {
-    furnitureId: '0',
-    tvId: '0',
-    simCardId: '0',
-    furnitureForm: {
-      buildingType: ['Free standing', 'House in complex', 'Townhouse', 'Apartment'],
-      parkingDistance: [15, 25, 50, 100],
-      deliveryFloor: ['Ground', '1', '2', '3+'],
-      liftStairs: ['Lift', 'Stairs']
-    },
-    TVorRICAMsg: 'You can\'t collect this order in store because your cart contains items which '
-      + 'require either RICA or TV License validation.',
-    MixedProductsMsg: 'We\'ll ship your furniture and other items in your cart to the selected address. '
-      + 'Only the furniture delivery fee will apply.'
-  };
-
   const checkCartCategories = () => {
+    const { items } = vtexjs.checkout.orderForm;
     const { categories } = CartController.state;
-    const allCategoriesIds = Object.keys(categories);
+    const { config } = CartController;
 
-    state.showFurnitureForm = allCategoriesIds.includes(config.furnitureId);
-    state.showTVIDForm = allCategoriesIds.includes(config.tvId);
-    state.showRICAMsg = allCategoriesIds.includes(config.simCardId);
+    state.showFurnitureForm = categories.includes(config.furnitureId);
+    state.showTVIDForm = categories.includes(config.tvId);
+    state.showRICAMsg = categories.includes(config.simCardId);
     state.showTVorRICAMsg = state.showTVIDForm || state.showRICAMsg;
+    /**
+      Conditions to show mixed products alert:
+      - more than one item
+      - after filter categories, this array includes at least one furniture id
+      - there are only one category OR not all the categories in the array are furniture
+    */
     state.showMixedProductsMsg = (
-      allCategoriesIds.includes(config.furnitureId)
-      && !allCategoriesIds.every((value) => value === config.furnitureId)
+      items.length > 1
+      && categories.includes(config.furnitureId)
+      && (categories.length === 1 || !categories.every((value) => value === config.furnitureId))
     );
   };
 
   const showCustomSections = () => {
+    const { config } = CartController;
+
     const tvRICAStepExists = ($('#tfg-custom-rica-msg').length > 0);
     const tvIDStepExists = ($('#tfg-custom-tvid-step').length > 0);
     const furnitureStepExists = ($('#tfg-custom-furniture-step').length > 0);
@@ -110,10 +104,8 @@ const ViewController = (() => {
         let tvCompleted = false;
 
         if (state.showFurnitureForm
-          && customShippingInfo.assembleFurniture
           && customShippingInfo.buildingType
           && customShippingInfo.deliveryFloor
-          && customShippingInfo.hasSufficientSpace
           && customShippingInfo.parkingDistance) {
           furnitureCompleted = true;
           validData = true;
@@ -135,11 +127,6 @@ const ViewController = (() => {
 
   const runCustomization = () => {
     if (window.location.hash === STEPS.SHIPPING || window.location.hash === STEPS.PAYMENT) {
-      if (typeof (setAppConfiguration) !== 'undefined') {
-        // eslint-disable-next-line no-undef
-        setAppConfiguration(config);
-      }
-
       setTimeout(() => {
         checkCartCategories();
 
@@ -156,6 +143,7 @@ const ViewController = (() => {
           if (!isDataCompleted) {
             setTimeout(async () => {
               isDataCompleted = await shippingCustomDataCompleted();
+
               if ((state.showFurnitureForm || state.showTVIDForm) && !isDataCompleted) {
                 window.location.hash = STEPS.SHIPPING;
               }