+++ CHANGELOG.md
@@ -19,3 +19,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Functionality associated with TV licence
 - Functionality associated with RICA
 - Functionality associated with Furniture products
+- Functionality associated with Mixed products